CORE-8751 Move GroupDetailsDialog show from presenter to view
