+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/ManageCollaboratorsView.java
@@ -2,7 +2,6 @@ package org.iplantc.de.collaborators.client;
 
 import org.iplantc.de.client.models.collaborators.Collaborator;
 import org.iplantc.de.client.models.groups.Group;
-import org.iplantc.de.collaborators.client.events.GroupNameSelected;
 import org.iplantc.de.collaborators.client.events.RemoveCollaboratorSelected;
 
 import com.google.gwt.resources.client.ImageResource;
@@ -17,8 +16,7 @@ import java.util.List;
  * 
  */
 public interface ManageCollaboratorsView extends IsWidget,
-                                                 RemoveCollaboratorSelected.HasRemoveCollaboratorSelectedHandlers,
-                                                 GroupNameSelected.HasGroupNameSelectedHandlers {
+                                                 RemoveCollaboratorSelected.HasRemoveCollaboratorSelectedHandlers {
 
     interface Appearance {
