+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/events/GroupNameSelected.java
@@ -4,7 +4,6 @@ import org.iplantc.de.client.models.groups.Group;
 
 import com.google.gwt.event.shared.EventHandler;
 import com.google.gwt.event.shared.GwtEvent;
-import com.google.gwt.event.shared.HandlerRegistration;
 
 /**
  * @author aramsey
@@ -13,10 +12,6 @@ public class GroupNameSelected extends GwtEvent<GroupNameSelected.GroupNameSelec
 
     public static interface GroupNameSelectedHandler extends EventHandler {
         void onGroupNameSelected(GroupNameSelected event);
-    }
-
-    public interface HasGroupNameSelectedHandlers {
-        HandlerRegistration addGroupNameSelectedHandler(GroupNameSelectedHandler handler);
     }
 
     public static Type<GroupNameSelectedHandler> TYPE = new Type<GroupNameSelectedHandler>();
